HollowGridMaker
Enter this code into a Java compiler and it will create a small grid with borders made of + and - for the top and bottom, and | for the sides.  inside a big square for you. You specify what the values are.

Uses nested loops 


Example:
For a grid size of 3 and cell size of 2, the output will be:

+--+--+--+
|     |     |    |
|     |     |    |
+--+--+--+
|     |     |    |
|     |     |    |
+--+--+--+
|     |     |    |
|     |     |    |
+--+--+--+